requirements.txt
torch>=2.0
numpy
scipy
matplotlib
pyyaml
h5py
scikit-learn
tqdm
